refactor: moved web3modal to web3 lib
